requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
streamlit

# LLM providers
openai>=1.0.0
google-generativeai